+++ app/components/layout/measureable-icon.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { observer } from 'mobx-react/native';
-import { Image } from 'react-native';
+import { Image, TouchableOpacity } from 'react-native';
 import vars from '../../styles/vars';
 import icons from '../helpers/icons';
 import beaconState from '../beacons/beacon-state';
@@ -8,13 +8,8 @@ import MeasureableView from '../shared/measureable-view';
 
 @observer
 export default class MeasureableIcon extends MeasureableView {
-    onMeasure = (position) => {
-        const { beacon, onPress, spotBgColor } = this.props;
-        if (!beacon) {
-            console.log('MeasureableIcon: ignoring empty beacon');
-            return;
-        }
-        // console.log(JSON.stringify(position));
+    request = (beacon, position) => {
+        const { onPress, spotBgColor } = this.props;
         beacon.position = position;
         beacon.content = this.content;
         beacon.onPressIcon = onPress;
@@ -22,14 +17,25 @@ export default class MeasureableIcon extends MeasureableView {
         beaconState.requestBeacon(beacon);
     };
 
-    // TODO clean up mock beacons
-    // ---------------------
+    onMeasure = (position) => {
+        const { beacon } = this.props;
+        if (!beacon) {
+            console.log('MeasureableIcon: ignoring empty beacon');
+            return;
+        }
+        if (Array.isArray(beacon)) {
+            beacon.forEach(b => this.request(b, position));
+        } else {
+            this.request(beacon, position);
+        }
+    };
+
     componentWillUnmount() {
         if (this.props.beacon) beaconState.removeBeacon(this.props.beacon.id);
     }
 
     get content() {
-        const { iconSource } = this.props;
+        const { iconSource, icon, color, testId } = this.props;
         if (iconSource) {
             return (
                 <Image
@@ -38,14 +44,20 @@ export default class MeasureableIcon extends MeasureableView {
                 />
             );
         }
-        return icons.plain(this.props.icon, undefined, this.props.color);
+        return icons.plain(icon, undefined, color, testId);
     }
 
     renderThrow() {
+        const { onPress } = this.props;
         return (
-            <MeasureableView onMeasure={this.props.beacon ? this.onMeasure : null}>
-                {this.content}
-            </MeasureableView>
+            <TouchableOpacity
+                pressRetentionOffset={vars.pressRetentionOffset}
+                onPress={onPress}>
+                <MeasureableView onMeasure={this.props.beacon ? this.onMeasure : null}>
+                    {this.content}
+                </MeasureableView>
+            </TouchableOpacity>
+
         );
     }
 }