+++ app/components/shared/view-receipts.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import _ from 'lodash';
-import { View, LayoutAnimation } from 'react-native';
+import { View } from 'react-native';
 import { reaction, observable, when } from 'mobx';
 import { observer } from 'mobx-react/native';
 import SafeComponent from '../shared/safe-component';
@@ -9,17 +9,25 @@ import ReadReceipt from './read-receipt';
 import { vars } from '../../styles/styles';
 import icons from '../helpers/icons';
 import Text from '../controls/custom-text';
+import { transitionAnimation } from '../helpers/animations';
 
 const receiptRow = {
     backgroundColor: vars.black25,
     alignSelf: 'flex-end',
     flexDirection: 'row',
     alignItems: 'center',
+    height: 20,
     width: 40,
-    marginLeft: 8,
-    marginRight: 8,
+    marginHorizontal: 8,
     borderRadius: 16,
     paddingHorizontal: 2
+};
+
+const emptyReceiptRow = {
+    alignSelf: 'flex-end',
+    height: 20,
+    width: 40,
+    marginHorizontal: 8
 };
 
 const half = {
@@ -41,7 +49,7 @@ export default class ViewReceipts extends SafeComponent {
         when(() => this.props.receipts && this.props.receipts.length, () => {
             this.prev = this.props.receipts.length;
             this.calculateLabel();
-            LayoutAnimation.easeInEaseOut();
+            transitionAnimation();
 
             this._observer = reaction(() => this.props.receipts && this.props.receipts.length, () => {
                 if (!this.props.receipts || !this.props.receipts.length) return;
@@ -91,7 +99,7 @@ export default class ViewReceipts extends SafeComponent {
 
     renderThrow() {
         const { receipts } = this.props;
-        if (!receipts || !receipts.length) return <View style={receiptRow} />;
+        if (!receipts || !receipts.length) return <View style={emptyReceiptRow} />;
 
         return (
             <View style={receiptRow}>