+++ app/components/signup/signup-step2.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { observable, action, reaction, when } from 'mobx';
 import { observer } from 'mobx-react/native';
-import { View, TouchableOpacity, LayoutAnimation } from 'react-native';
+import { View, TouchableOpacity } from 'react-native';
 import randomWords from 'random-words';
 import Text from '../controls/custom-text';
 import { vars, signupStyles } from '../../styles/styles';
@@ -14,8 +14,8 @@ import buttons from '../helpers/buttons';
 import SignupButtonBack from './signup-button-back';
 import SignupHeading from './signup-heading';
 import SignupStepIndicator from './signup-step-indicator';
-import TmHelper from '../../telemetry/helpers';
 import tm from '../../telemetry';
+import { transitionAnimation } from '../helpers/animations';
 
 const { S } = telemetry;
 
@@ -23,6 +23,14 @@ const { validators } = validation;
 const { username } = validators;
 
 const MAX_USERNAME_LENGTH = config.user.maxUsernameLength;
+
+const sublocation = S.ACCOUNT_USERNAME;
+
+const tmUsername = {
+    item: S.USERNAME,
+    location: S.ONBOARDING,
+    sublocation
+};
 
 const suggestionContainerHeight = signupStyles.suggestionContainer.maxHeight;
 
@@ -37,7 +45,6 @@ export default class SignupStep2 extends SafeComponent {
 
     componentDidMount() {
         this.startTime = Date.now();
-        TmHelper.currentRoute = S.ACCOUNT_USERNAME;
         // QUICK SIGNUP DEV FLAG
         if (__DEV__ && process.env.PEERIO_QUICK_SIGNUP) {
             this.usernameInput.onChangeText(randomWords({ min: 2, max: 2, join: 'o' }).substring(0, 16));
@@ -50,7 +57,7 @@ export default class SignupStep2 extends SafeComponent {
 
         this.suggestionAnimationReaction = reaction(
             () => signupState.usernameSuggestions,
-            () => LayoutAnimation.easeInEaseOut()
+            transitionAnimation
         );
         signupState.suggestUsernames();
 
@@ -61,14 +68,14 @@ export default class SignupStep2 extends SafeComponent {
 
     componentWillUnmount() {
         this.suggestionAnimationReaction();
-        tm.signup.duration(this.startTime);
+        tm.signup.duration({ sublocation, startTime: this.startTime });
     }
 
     @action.bound handleNextButton() {
         if (this.isNextDisabled) return;
         signupState.username = this.usernameState.value;
         signupState.next();
-        tm.signup.navigate(S.NEXT);
+        tm.signup.navigate({ sublocation, option: S.NEXT });
     }
 
     get isNextDisabled() { return !socket.connected || !this.usernameState.value || !this.usernameInput.isValid; }
@@ -145,13 +152,13 @@ export default class SignupStep2 extends SafeComponent {
             <View style={signupStyles.page}>
                 <SignupStepIndicator />
                 <View style={signupStyles.container}>
-                    <SignupButtonBack />
+                    <SignupButtonBack telemetry={{ sublocation, option: S.BACK }} />
                     <SignupHeading title="title_createYourAccount" subTitle="title_usernameHeading" />
                     <StyledTextInput
                         autoFocus
                         state={this.usernameState}
                         validations={username}
-                        inputName={S.USERNAME}
+                        telemetry={tmUsername}
                         helperText={this.usernameState.value.length >= MAX_USERNAME_LENGTH ?
                             tx('title_characterLimitReached') :
                             tx('title_hintUsername')}